Standardize block part card
